+++ apps/opplaringspenger-soknad/src/app/søknad/steps/kurs/kursMessages.ts
@@ -38,6 +38,7 @@ const nb = {
         'For mange tegn. Beskrivelsen kan ikke bestå av flere enn {antall} tegn.',
     'steg.kurs.validation.reisedager.reisedagUtenforKursperiode':
         '{antallDager, plural, one {Én reisedag} other {# reisedager}} er utenfor søknadsperioden: {dager}. Dette kan skje hvis du har endret hvilke dager du søker opplæringspenger, etter at du har lagt til reisedag. Hvis du vil søke om opplæringspenger for disse dagene, må du inkludere dem i perioden du søker om opplæringspenger. Hvis du ikke søker opplæringspenger disse dagene, må du fjerne dem her.',
+    'steg.kurs.validation.reisedager.reisedagPåHelg': 'En reisedag kan ikke være en lørdag eller søndag.',
     'steg.kurs.validation.ferieuttak.ferieperiodeUtenforKursperiode':
         'Ferie er ikke innenfor dagene du søker. Dette kan skje hvis du har endret hvilke dager du søker opplæringspenger, etter at du har lagt til ferie. Du må endre eller fjerne ferie som ikke er innenfor dager du søker. Feriedager utenfor søknadsperiode: {dager}.',
     'steg.kurs.validation.opplæringsinstitusjon.stringHasNoValue': 'Du må fylle ut hvor opplæringen foregår.',
@@ -128,6 +129,7 @@ const nn: Record<keyof typeof nb, string> = {
         'For få teikn. Du må beskrive kvifor du reiser på andre dagar enn du har kurs eller opplæring med fleire teikn.',
     'steg.kurs.validation.reisedagerBeskrivelse.stringIsTooLong':
         'For mange teikn. Beskrivinga kan ikkje bestå av fleire enn {antall} teikn.',
+    'steg.kurs.validation.reisedager.reisedagPåHelg': 'Ein reisedag kan ikkje vera ein laurdag eller sundag.',
     'steg.kurs.validation.reisedager.reisedagUtenforKursperiode':
         'Reisedag er ikkje innanfor dagane du søkjer. Dette kan skje om du har endra kva dagar du søkjer opplæringspengar, etter at du har lagt til reisedag. Du må endre eller fjerne reisedag som ikkje er innanfor dagane du søkjer.',
     'steg.kurs.validation.ferieuttak.ferieperiodeUtenforKursperiode':