+++ apps/ungdomsytelse-deltaker/mock/utils/mockUtils.ts
@@ -1,5 +1,3 @@
-// import { OppgaveStatus } from '@navikt/ung-common';
-
 import { UngdomsytelseInntektsrapportering } from '@navikt/k9-brukerdialog-prosessering-api';
 import { OppgaveDto } from '@navikt/ung-deltakelse-opplyser-api-deltaker';