+++ apps/ungdomsytelse-deltaker/playwright/tests/test.spec.ts
@@ -1,0 +1,6 @@
+import { expect, test } from '@playwright/test';
+
+test('Fyll ut søknad og kontroller oppsummering', async ({ page }) => {
+    console.log(page);
+    expect(true).toBe(true);
+});