+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/modules/oppgaver/rapporter-inntekt/RapporterInntektOppgavePage.tsx
@@ -1,8 +1,9 @@
-import DefaultPageLayout from '@innsyn/pages/layout/DefaultPageLayout';
 import { getOppgaveDokumentTittel } from '@innsyn/utils/textUtils';
+import { VStack } from '@navikt/ds-react';
 import { OppgaveStatus } from '@navikt/ung-deltakelse-opplyser-api-deltaker';
 import { commonQueries } from '@shared/api/queries/commonQueries';
 import { useAppIntl } from '@shared/i18n';
+import DefaultPageLayout from '@shared/pages/layout/DefaultPageLayout';
 import { RapporterInntektOppgave } from '@shared/types/Oppgave';
 import { useQueryClient } from '@tanstack/react-query';
 import { useEffect } from 'react';
@@ -37,7 +38,9 @@ const RapporterInntektOppgavePage = (props: RapporterInntektOppgaveProps) => {
 
     return (
         <DefaultPageLayout documentTitle={getOppgaveDokumentTittel(props.oppgave, appIntl)}>
-            <RapporterInntektOppgavePart {...props} />
+            <VStack gap="6">
+                <RapporterInntektOppgavePart {...props} />
+            </VStack>
         </DefaultPageLayout>
     );
 };