+++ apps/ungdomsytelse-deltaker/src/apps/søknad/steg/barn/BarnInfo.tsx
@@ -1,7 +1,7 @@
 import { BodyShort } from '@navikt/ds-react';
 import { RegistrertBarn } from '@navikt/sif-common-api';
 
-import { AppText } from '../../../../i18n';
+import { AppText } from '@shared/i18n';
 import BarnList from './BarnList';
 
 interface Props {