+++ packages/sif-common-hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-common-hooks
 
+## 0.3.6
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 0.3.5
 
 ### Patch Changes