+++ packages/sif-common-sentry/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @navikt/sif-common-sentry
 
+## 0.28.42
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.28.41
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.28.40
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.28.39
+
+### Patch Changes
+
+- 52d9b9e: Pakkeoppdateringer - minor/patch
+
+## 0.28.38
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.28.37
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch. Fjerne ubrukte pakker.
+
+## 0.28.36
+
+### Patch Changes
+
+- Pakkeoppdataringer - minor/patch
+
 ## 0.28.35
 
 ### Patch Changes